Starting to work on separating out settings and adding to docs
