+++ .docfx/manual/features/serializable-dictionary.md
@@ -1,0 +1,3 @@
+# Serializable Dictionary
+
+One of the most requested features for Unity is the ability to serialize dictionary types in the editor for use at runtime. With the release of `Unity 2020.1` a required supporting feature landed allowing for template based property drawers.
