+++ .docfx/manual/features/visual-scripting.md
@@ -1,0 +1,2 @@
+# Visual Scripting Support
+
