+++ GDX/Editor/ProjectSettings.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 6ec3c82170fe464098104b561100f732
+timeCreated: 1616763060
