+++ GDX/Editor/ProjectSettings/AutomaticUpdatesSettings.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 6d688688129a4ec2b8942725b885a839
+timeCreated: 1616763169
