+++ GDX/Editor/ProjectSettings/BuildInfoSettings.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 460158ee794846339050e132093e8932
+timeCreated: 1616763217
