+++ GDX/Editor/ProjectSettings/EnvironmentSettings.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: deeeb500cc9b4acca27b3c2e55b7b2af
+timeCreated: 1616763243
