+++ GDX/Editor/ProjectSettings/LocaleSettings.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: be585a867a5c4575bdee8d709a7e01f3
+timeCreated: 1616763256
