+++ GDX/Editor/ProjectSettings/SettingsProvider.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: ad76b3aead804dda80ce6b82014cc7dc
+timeCreated: 1616763124
