+++ GDX/Editor/ProjectSettings/VisualScriptingSettings.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 76dd6d94db51bce4db47efa9d05a0e69
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
