--- GDX/Editor/Settings.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: f96e8fc62ba248a389f5c4e35db13666
-timeCreated: 1610682056
