+++ GDX/Runtime/ByteExtensions.cs
@@ -6,7 +6,7 @@ namespace GDX
     /// <summary>
     ///     <see cref="byte" /> Based Extension Methods
     /// </summary>
-    [VisualScripting(VisualScriptingAttribute.Category.Extensions)]
+    [VisualScriptingNode(VisualScriptingNodeAttribute.Category.Extensions)]
     public static class ByteExtensions
     {
         /// <summary>
