+++ GDX/Runtime/Collections/Pooling/GameObjectPool.cs
@@ -10,7 +10,7 @@ namespace GDX.Collections.Pooling
     ///     <see cref="GameObject" /> based functionality extending the <see cref="ListManagedPool" /> to better support
     ///     <see cref="GameObject" /> patterns.
     /// </summary>
-    [VisualScripting(VisualScriptingAttribute.Category.Types)]
+    [VisualScriptingNode(VisualScriptingNodeAttribute.Category.Types)]
     public static class GameObjectPool
     {
         /// <summary>
