+++ GDX/Runtime/Collections/Pooling/ManagedPools.cs
@@ -10,7 +10,7 @@ namespace GDX.Collections.Pooling
     ///     A managed pooling system implementation primarily meant for the object oriented patterns, based on the C# base
     ///     object.
     /// </summary>
-    [VisualScripting(VisualScriptingAttribute.Category.Extensions)]
+    [VisualScriptingNode(VisualScriptingNodeAttribute.Category.Extensions)]
     public static class ManagedPools
     {
         /// <summary>
