+++ GDX/Runtime/GDX.asmdef
@@ -34,6 +34,11 @@
             "name": "com.unity.platforms",
             "expression": "0.9",
             "define": "GDX_PLATFORMS"
+        },
+        {
+            "name": "com.unity.visualscripting",
+            "expression": "1.5.2",
+            "define": "GDX_VISUALSCRIPTING"
         }
     ],
     "noEngineReferences": false
