+++ GDX/Runtime/GameObjectExtensions.cs
@@ -12,7 +12,7 @@ namespace GDX
     /// <remarks>
     ///     Used for MonoBehaviour workflows.
     /// </remarks>
-    [VisualScripting(VisualScriptingAttribute.Category.Extensions)]
+    [VisualScriptingNode(VisualScriptingNodeAttribute.Category.Extensions)]
     public static class GameObjectExtensions
     {
         /// <summary>
