+++ GDX/Runtime/Mathematics/Random/MersenneTwister.cs
@@ -24,7 +24,7 @@ namespace GDX.Mathematics.Random
     ///         for details on the algorithm.
     ///     </para>
     /// </remarks>
-    [VisualScripting(VisualScriptingAttribute.Category.Types)]
+    [VisualScriptingNode(VisualScriptingNodeAttribute.Category.Types)]
     public class MersenneTwister : System.Random
     {
         /// <summary>
