+++ GDX/Runtime/MeshExtensions.cs
@@ -11,7 +11,7 @@ namespace GDX
     /// <summary>
     ///     <see cref="UnityEngine.Mesh" /> Based Extension Methods
     /// </summary>
-    [VisualScripting(VisualScriptingAttribute.Category.Extensions)]
+    [VisualScriptingNode(VisualScriptingNodeAttribute.Category.Extensions)]
     public static class MeshExtensions
     {
         /// <summary>
