+++ GDX/Runtime/Vector2Extensions.cs
@@ -13,7 +13,7 @@ namespace GDX
     /// <summary>
     ///     <see cref="Vector2" /> Based Extension Methods
     /// </summary>
-    [VisualScripting(VisualScriptingAttribute.Category.Extensions)]
+    [VisualScriptingNode(VisualScriptingNodeAttribute.Category.Extensions)]
     public static class Vector2Extensions
     {
         /// <summary>
